mapquest request works

+++ src/main/java/at/technikumwien/tourplanner_backend/repository/TourRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface TourRepository
-        extends JpaRepository<Tour, Integer> {
+        extends JpaRepository<Tour, Long> {
 }
